+++ packages/people-and-teams/profilecard/src/components/IconLabel.tsx
@@ -1,0 +1,50 @@
+import * as React from 'react';
+import IconLocation from '@atlaskit/icon/glyph/location';
+import IconRecent from '@atlaskit/icon/glyph/recent';
+import IconMention from '@atlaskit/icon/glyph/mention';
+import IconEmail from '@atlaskit/icon/glyph/email';
+import OfficeBuildingIcon from '@atlaskit/icon/glyph/office-building';
+
+import {
+  DetailsLabel,
+  DetailsLabelIcon,
+  DetailsLabelText,
+} from '../styled/Card';
+
+const icons = {
+  location: IconLocation,
+  time: IconRecent,
+  mention: IconMention,
+  email: IconEmail,
+  companyName: OfficeBuildingIcon,
+};
+
+type Props = {
+  icon: string;
+  children?: React.ReactNode;
+};
+
+export default class IconLabel extends React.PureComponent<Props> {
+  static defaultProps = {
+    icon: '',
+  };
+
+  render() {
+    if (!this.props.children) {
+      return null;
+    }
+
+    // @ts-ignore
+    const IconElement = this.props.icon && icons[this.props.icon];
+    const displayIcon = IconElement ? (
+      <IconElement label={`icon ${this.props.icon}`} size="small" />
+    ) : null;
+
+    return (
+      <DetailsLabel>
+        <DetailsLabelIcon>{displayIcon}</DetailsLabelIcon>
+        <DetailsLabelText>{this.props.children}</DetailsLabelText>
+      </DetailsLabel>
+    );
+  }
+}