+++ packages/elements/mention/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/mention
 
+## 17.6.3
+- [patch] [14b89652d7](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/14b89652d7):
+
+  - TEAMS-364 : Renaming the missed test files in the previous PR
+
 ## 17.6.2
 - [patch] [c8b669eaec](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c8b669eaec):